lambdas without provider
