+++ cdk/domino_cdk/provisioners/lambda_files/cluster_post_deletion_tasks.py
@@ -1,17 +1,22 @@
+import json
 import os
 
 import boto3
+import requests
 
 
 def on_event(event, context):
+    print('Debug: event: ', event)
+    print('Debug: environ:', os.environ)
     request_type = event['RequestType']
+    response = {}
     if request_type == 'Create':
-        return on_create(event)
-    if request_type == 'Update':
-        return None
+        response = on_create(event)
     if request_type == 'Delete':
-        return on_delete(event)
-    raise Exception('Invalid request type: %s' % request_type)
+        response = on_delete(event)
+    event.update(response)
+    event['Status'] = 'SUCCESS'
+    requests.put(event['ResponseURL'], data=json.dumps(event))
 
 
 def on_create(event):
@@ -29,3 +34,4 @@ def on_delete(event):
     for lg in response['logGroups']:
         print(f'Change retention of log group {lg["logGroupName"]}')
         logs_client.put_retention_policy(logGroupName=lg['logGroupName'], retentionInDays=1)
+    return {}
